pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "python-dotenv>=1.1.1",
]
